Refactoring: GeoDataRiverSurvey::RotateRiverCrosssectionCommand is moved to separated files.
